backend/requirements.txt
Authlib==1.3.2
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.0.1
Werkzeug==3.1.3
